Fixed a bug where the response buffer wasn't being cleared on auth failure.
If a request was retried due to an auth failure, the previous response
(ie the response payload from the auth failure) would not be cleared
from the response writer.  This would cause the successful retry to
return both the error response and the successful response, concatenated
together.  This then causes anything trying to parse the response to
parse the failure JSON response instead of the successful response.